requirements.txt
fastapi
uvicorn
ollama
streamlit
faiss-cpu
transformers
torch
PyPDF2
pypdf
requests 
langchain
langchain-community
langchain-huggingface
langchain-core
langchain-text-splitters
tiktoken
scikit-learn
sentence-transformers
regex
plotly
spacy
networkx
matplotlib